requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.7.1
execnet==2.1.1
Faker==37.3.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
h11==0.14.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.37.13
locust-cloud==1.26.2
MarkupSafe==3.0.2
msgpack==1.1.1
mysql-connector-python==9.3.0
outcome==1.3.0.post0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
psutil==7.0.0
pycparser==2.22
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
pytest-xdist==3.8.0
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pywin32==311
pyzmq==27.0.0
requests==2.32.4
selenium==4.30.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.1
zope.interface==7.2
